feat: changed sig/pig create, update logic about the content
