+++ src/model/sig.py
@@ -24,7 +24,7 @@ class SIG(SQLModel, table=True):
 
     title: str = Field(nullable=False)
     description: str = Field(nullable=False)
-    content_src: str = Field(nullable=False, unique=True)
+    content_id: Optional[int] = Field(foreign_key="article.id", unique=True)
 
     status: SIGStatus = Field(nullable=False)
