+++ src/routes/pig.py
@@ -1,4 +1,4 @@
-from typing import Optional, Sequence
+from typing import Sequence
 
 from fastapi import APIRouter, Depends, HTTPException
 from pydantic import BaseModel
@@ -8,51 +8,16 @@ from sqlmodel import select
 from src.auth import get_current_user
 from src.db import SessionDep
 from src.model import PIG, PIGGlobalStatus, PIGMember, PIGStatus, User, UserRole
-from src.util import is_valid_semester, is_valid_year
 from src.util import get_pig_global_status
+from src.controller import create_pig_controller, BodyCreatePIG, update_pig_controller, BodyUpdatePIG
+
 
 pig_router = APIRouter(tags=['pig'])
 
 
-class BodyCreatePIG(BaseModel):
-    title: str
-    description: str
-    content_src: str
-    year: int
-    semester: int
-
-
 @pig_router.post('/pig/create', status_code=201)
 async def create_pig(body: BodyCreatePIG, session: SessionDep, current_user: User = Depends(get_current_user), pig_global_status: PIGGlobalStatus = Depends(get_pig_global_status)) -> PIG:
-    if pig_global_status.status != PIGStatus.surveying: raise HTTPException(400, "cannot create pig when pig global status is not surveying")
-    if not is_valid_year(body.year): raise HTTPException(422, detail="invalid year")
-    if not is_valid_semester(body.semester): raise HTTPException(422, detail="invalid semester")
-
-    pig = PIG(
-        title=body.title,
-        description=body.description,
-        content_src=body.content_src,
-        year=body.year,
-        semester=body.semester,
-        owner=current_user.id,
-        status=PIGStatus.surveying
-    )
-    session.add(pig)
-    try: session.commit()
-    except IntegrityError:
-        session.rollback()
-        raise HTTPException(409, detail="unique field already exists")
-    session.refresh(pig)
-    if pig.id is None: raise HTTPException(503, detail="pig primary key does not exist")
-    pig_member = PIGMember(
-        ig_id=pig.id,
-        user_id=current_user.id,
-        status=pig.status
-    )
-    session.add(pig_member)
-    session.commit()
-    session.refresh(pig)
-    return pig
+    return await create_pig_controller(body, session, current_user, pig_global_status)
 
 
 @pig_router.get('/pig/{id}')
@@ -67,34 +32,9 @@ async def get_all_pigs(session: SessionDep) -> Sequence[PIG]:
     return session.exec(select(PIG)).all()
 
 
-class BodyUpdateMyPIG(BaseModel):
-    title: str
-    description: str
-    content_src: str
-    year: int
-    semester: int
-
-
 @pig_router.post('/pig/{id}/update', status_code=204)
-async def update_my_pig(id: int, body: BodyUpdateMyPIG, session: SessionDep, current_user: User = Depends(get_current_user)) -> None:
-    pig = session.get(PIG, id)
-    if not pig: raise HTTPException(404, detail="pig not found")
-    if pig.owner != current_user.id: raise HTTPException(
-        status_code=403, detail="cannot update pig of other")
-    if not is_valid_year(body.year): raise HTTPException(422, detail="invalid year")
-    if not is_valid_semester(body.semester): raise HTTPException(422, detail="invalid semester")
-
-    pig.title = body.title
-    pig.description = body.description
-    pig.content_src = body.content_src
-    pig.year = body.year
-    pig.semester = body.semester
-    session.add(pig)
-    try: session.commit()
-    except IntegrityError:
-        session.rollback()
-        raise HTTPException(409, detail="unique field already exists")
-    return
+async def update_my_pig(id: int, body: BodyUpdatePIG, session: SessionDep, current_user: User = Depends(get_current_user)) -> None:
+    await update_pig_controller(id, body, session, current_user)
 
 
 @pig_router.post('/pig/{id}/delete', status_code=204)
@@ -107,39 +47,9 @@ async def delete_my_pig(id: int, session: SessionDep, current_user: User = Depen
     return
 
 
-class BodyUpdatePIG(BaseModel):
-    title: Optional[str] = None
-    description: Optional[str] = None
-    content_src: Optional[str] = None
-    status: Optional[PIGStatus] = None
-    year: Optional[int] = None
-    semester: Optional[int] = None
-
-
 @pig_router.post('/executive/pig/{id}/update', status_code=204)
 async def update_pig(id: int, body: BodyUpdatePIG, session: SessionDep) -> None:
-    pig = session.get(PIG, id)
-    if pig is None: raise HTTPException(404, detail="no pig exists")
-    if body.title:
-        pig.title = body.title
-    if body.description:
-        pig.description = body.description
-    if body.content_src:
-        pig.content_src = body.content_src
-    if body.status:
-        pig.status = body.status
-    if body.year:
-        if not is_valid_year(body.year): raise HTTPException(422, detail="invalid year")
-        pig.year = body.year
-    if body.semester:
-        if not is_valid_semester(body.semester): raise HTTPException(422, detail="invalid semester")
-        pig.semester = body.semester
-    session.add(pig)
-    try: session.commit()
-    except IntegrityError:
-        session.rollback()
-        raise HTTPException(409, detail="unique field already exists")
-    return
+    await update_pig_controller(id, body, session, None)
 
 
 @pig_router.post('/executive/pig/{id}/delete', status_code=204)
